+++ G03_CENG431_Final/src/view/LabelInfo.java
@@ -4,12 +4,12 @@ import java.awt.Color;
 
 public class LabelInfo {
 	public Color color;
-	public Float percent;
+	public Double percent;
 	public String coinName;
 	public Float value;
 	public String banknote;
 
-	public LabelInfo(Color color, Float percent, String name, Float value, String banknote) {
+	public LabelInfo(Color color, Double percent, String name, Float value, String banknote) {
 		this.color = color;
 		this.percent = percent;
 		this.coinName = name;
